Move _mpl_cmap2pgf_cmap to _color.py
